requirements.txt
selenium>=3.141,<3.142
slackclient>=2.2,<2.3
